refactor: root renderer
