+++ .mcp.json.example
@@ -2,7 +2,7 @@
   "mcpServers": {
     "browser-tools": {
       "command": "npx",
-      "args": ["-y", "@modelcontextprotocol/server-browser-tools"]
+      "args": ["-y", "@agentdeskai/browser-tools-mcp@1.2.0"]
     },
     "context7": {
       "type": "sse",
