+++ libs/frontend/abstract/application/src/renderer/root-renderer.interface.ts
@@ -1,6 +1,12 @@
 import type { IRendererModel } from './renderer.model.interface'
 
-export type IRootRenderer = React.JSXElementConstructor<{
-  ref?: React.RefObject<HTMLDivElement | null>
+export interface IRootRendererProps {
+  containerStyle?: React.CSSProperties
   renderer: IRendererModel
-}>
+}
+
+export type IRootRenderer = React.JSXElementConstructor<
+  {
+    ref?: React.RefObject<HTMLDivElement | null>
+  } & IRootRendererProps
+>
